Fix refactor by reordering imports to mock first
